requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

Cython
librosa
matplotlib
numpy
phonemizer
scipy
tensorboard
torch        
torchaudio  
Unidecode
amfm_decompy
openai-whisper
rich
gradio
jieba
cn2an
transformers
pypinyin